ai-system/requirements.txt
# Core AI libraries
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
nltk>=3.8.0
spacy>=3.6.0

# Database
mysql-connector-python>=8.0.0
pymysql>=1.0.0
sqlite3

# Web framework
flask>=2.3.0
flask-cors>=4.0.0

# Vanna Agent
vanna>=0.0.1

# Text processing
jieba>=0.42.1
underthesea>=1.3.0

# Model training
joblib>=1.3.0
pickle-mixin>=1.0.2

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Chatbot & RAG Dependencies
sentence-transformers
transformers
torch
faiss-cpu
langchain
langchain-community
